Add files via upload
The circuit diagram contains all the key elements of the pH probe circuit, including the offset voltage potential divider circuit and the pH measuring circuit.
To combat fluctuation errors, we tried adding the offset voltage as an extra input into the microcontroller, instead of setting a constant offset voltage.
Initially, instead of a potentiometer, the design had a fixed potential divider, consisting of resistors with 11kΩ and a 1kΩ. This should have theoretically supplied an offset voltage of 500mV, however, due to the tolerances in the resistors, the actual offset was only 430mV. By replacing the potential divider with a potentiometer, we expected to get a better accuracy of the offset. 
The buffer operational amplifier is essential for the voltage offset circuit as it maintains the voltage level of the reference probe at the same level as the output of the potential divider.